==> 1_prep_collar_data.Rmd <==
---
title: "1. Pre-process Collar Data"
author: "Amy Van Scoyoc"
date: "11/21/2022"
output: github_document
---

# Purpose 
This document is the code to format the coyote GPS data into a single dataframe. Coyote data were downloaded from the Vectronic Inventa website and named based on first and last date in the file. The date key file contains the field-season, the coyote release date-times, and the collar drop date-times. 

## Load libraries

```{r, warning=F, message=F}
library(here) #reproducible working directory
library(tidyverse) #for data wrangling
library(lubridate) #for data/time
library(sp) #for UTM transformation
```

### Import and Combine Coyote field seasons
```{r warning=F, message=F}
#data table with coyote sex, id, collar start_date, collar end_date
datekey <- read_csv(here("data","coyote_datekey.csv")) 

#season 1 data (8-weeks)
season1 <- read_csv(here("data", "coyote_2020.11.08_2021.06.30.csv")) %>% mutate(season = 1)

#season 2 data (16-weeks)
season2 <- read_csv(here("data", "coyote_2021.12.01_2022.03.04.csv")) %>% mutate(season = 2)

#season 3 data (10.5-weeks)
season3 <- read_csv(here("data", "coyote_2022.11.27_2023.02.10.csv")) %>% mutate(season = 3)

#fix column names
names(season3)=names(season2)

#combine field seasons
dat_all <- rbind(season1, season2) %>% rbind(season3) %>% 
  dplyr::select(x = `Longitude[deg]`, y = `Latitude[deg]`, 
                t = `Acq. Time [UTC]`, collar_id = `Collar ID`, season) %>%
  mutate(t = lubridate::with_tz(ymd_hms(t,tz="UTC"),"America/Los_Angeles")) %>% #adjust time
  drop_na(x) %>% #drop any missing locations
  left_join(., datekey, by = c("season", "collar_id")) %>% #join date table for filtering
  filter(t >= start_3d & t <= end, #filter 3-days after collaring
         ID != "M4") #remove, M4 died 5-days after release via vehicle collision
```
### Calculate GPS data metrics

There were hourly GPS locations for a total of 190 unique days (mean 60 days per individual) from November to May of 2020-2021 and 2021-2022. For three seaons of data there were 265 unique days (mean 60 days per individual).
```{r}
#calculate number of collared coyote dates (190 unique days)
dat_all %>% 
  mutate(julian_day = yday(t)) %>% 
  group_by(season) %>% 
  distinct(julian_day) %>% 
  nrow() #count unique julian day by season
```

```{r}
#calculate mean days across individuals (mean 54 days)
dat_all %>% 
  group_by(ID) %>% 
  summarize(start = min(as.Date(t)), #first date per individual
            end = max(as.Date(t)), #last date per individual
            ndays = difftime(end, start, units = "days"), #n days per individual
            npoints = n()) %>% #number points
  summarize(n = mean(ndays)) #mean n days per individual
```

### Transform to UTMs
```{r}
#transform coordinates
llcoord <- SpatialPoints(dat_all[,1:2],proj4string=CRS("+init=epsg:4326")) #transform to sp object
utmcoord <- spTransform(llcoord, CRS("+init=epsg:26910")) #transform to utms

# overwrite UTM locations to data frame
dat_all$x <- attr(utmcoord,"coords")[,1]
dat_all$y <- attr(utmcoord,"coords")[,2]
```

### Export dataset to object for later use
```{r}
#arrange and convert to dataframe necessary for moveHMM prepData function
dat_all <- dat_all %>% 
  dplyr::select(ID, x, y, t) %>% #select columns
  arrange(ID, t) %>% #arrange by ID and timestamp
  data.frame() #convert to dataframe

save(dat_all, file = "data_objects/dat_all.Rdata") #save object
```
