==> 2_prep_covariates.Rmd <==
---
title: "covariates"
author: "Amy Van Scoyoc"
date: "4/15/2022"
output: github_document
---

### Import Libraries
```{r}
library(here) #reproducible working directory
library(tidyverse) #data tidying
library(sf) #modifying spatial attributes 
library(raster) #reclassify, stack environmental covariates
library(sdmvspecies) #rescale environmental covariates 
library(tmap) #plots
```

### Reclassify Land Cover

```{r}
#import landcover layer
lc <- raster(here("data_covariates", "covariates_from_GEE", "HREC_landcover_extended_2016.tiff"))
#import elevation layer
elev <- raster(here("data_covariates", "covariates_from_GEE", "HREC_elevation_extended.tif"))
#create ruggedness (Terrain Ruggedness Index) layer
tri <- raster::terrain(elev, "TRI", neigbors=8, 
                       filename = here("data_covariates", "HREC_tri.tif"), overwrite=T)

#import Hopland Research and Extension Center boundary  
hopland <- st_read(here("data_covariates", "HREC_boundary", "HREC_boundary.shp"))
#set crs HREC boundary 
hopland <- st_transform(hopland, st_crs(lc))
```

```{r}
#create development vector
dev_vec <- c(-1, 21, NA, #set others to NA 
            21, 25, 1,  #combine developed landcover 22, 23, 24
            25, Inf, NA) %>% #set others to NA
            matrix(., ncol = 3, byrow = T)

#create forest road vector
road_vec <- c(-1, 20, NA, #set others to NA 
            20, 21, 1,  #combine light dev landcover 21
            21, Inf, NA) %>% #set others to NA
            matrix(., ncol = 3, byrow = T)

#create cropland layer
crop_vec <- c(-1, 80, NA, #set others to NA 
            80, 82, 1,  #combine grass pasture layer 81, 82
            82, Inf, NA) %>% #set others to NA
            matrix(., ncol = 3, byrow = T)

#habitat layer
veg_vec <- c(-1, 40, NA, #set others to NA 
            40, 43, 2,  #combine forest 41, 42, 43 (med visibility)
            43, 51, NA, #set others to NA 
            51, 52, 1, #combine shrub 52 (least visibility)
            52, 70, NA, #set others to NA  
            70, 71, 3, #combine grassland 71 (most visible) 
            71, Inf, NA) %>% #set others to NA  
            matrix(., ncol = 3, byrow = T)
```

```{r}
#reclassify land cover values 
dev <- reclassify(lc, dev_vec)
road <- reclassify(lc, road_vec)
crop <- reclassify(lc, crop_vec)
veg <- reclassify(lc, veg_vec)

#convert vegetation raster to categorical values (For lions)
type <- data.frame(ID=1:3, cover=c("Shrubland", "Woodland", "Grassland"))
levels(veg) <- type
```

```{r}
#create distance rasters
dev_dist <- gridDistance(dev, origin = 1)
road_dist <- gridDistance(road, origin = 1)
crop_dist <- gridDistance(crop, origin = 1)
```

### Hunter risk layer

Model = Habitat + Road Dist + Rugged + (Road Dist * Habitat) 
*(Gaynor et al. 2021)*

```{r}
#Create fake data
set.seed(42)
dummy_data <- data.frame(used = rbinom(30, 1, 0.4),
                         vegetation = factor(rep(c("Shrubland", "Woodland", "Grassland"), each = 10), 
                                              levels = c("Shrubland", "Woodland", "Grassland")),
                         road_dist = c(runif(30, min = 0, max = 2500)),
                         ruggedness = c(runif(30, min = 0, max = 300))) 

#Recreate formula from Gaynor et. al 2021 and specify interactions:
mod <- glm(formula = used ~ vegetation + road_dist + ruggedness + road_dist*vegetation, 
                  data = dummy_data, family = binomial(link = "logit"))

#Create a new model to update the coefficients:
mod_real <- mod

#Manually create vector of REAL coefficients from Gaynor et al. 2021:
mod_real$coefficients <-c(-3.73, 2.08, 2.75, -2.23, 0.15, 1.53, 1.74)
```


```{r}
#Create hunter risk layer from Gaynor et al. 2021 model

#rescale environmental layers 
road_sc <- sdmvspecies::rescale(road_dist)
tri_sc <- sdmvspecies::rescale(tri)

#stack layers
hunt_stack <- brick(veg, road_sc, tri_sc) 
names(hunt_stack) <- c("vegetation", "road_dist", "ruggedness")
#plot(hunt_stack, nc = 3)

#convert to matrix 
hunt_values <- values(hunt_stack)
head(hunt_values)

#vegetation class key
veg_table <- unique(data.frame(vegetation = c(1,2,3),
                               vegetation_class = as.factor(c("Shrubland","Woodland","Grassland"))))

#create dataframe with categorical values for prediction
hunt_vals <- hunt_values %>% data.frame %>% 
  mutate(rid = 1:length(vegetation), nick_name = "dummy") %>% 
  merge(veg_table, all.x=TRUE) %>% #merge categorical names
  dplyr::select(-vegetation) %>% #drop numeric
  rename(vegetation = vegetation_class) %>% #keep factor class
  arrange(rid) #add row ID
str(hunt_vals)

#predict values
hunter_prediction <- predict(mod_real, newdata = hunt_vals, allow.new.levels=TRUE)

#fill raster with values
hunter_risk <- hunt_stack[[3]] %>% setValues(hunter_prediction)

plot(hunter_risk)
```

```{r}
hunt_rsf <- exp(hunter_risk)/max(values(exp(hunter_risk)), na.rm = TRUE)
(breaks <- quantile(hunter_prediction, seq(0,1,1/10), na.rm = TRUE))
hunter_discrete <- cut(hunter_prediction, breaks = breaks, label = 1:10)
hunter <- hunt_rsf %>% setValues(hunter_discrete)

#create mask for where hunters won't hunt (150m of development)
dev_mask <- raster::buffer(dev, width = 150)
hunter_risk <- mask(hunter, dev_mask, inverse=TRUE) #remove dev from hunter risk

# replacing NA's by zero
hunter_risk[is.na(hunter_risk[])] <- 0 

plot(hunter_risk)
```


```{r}
#crop to hopland boundary to see if its accurate
# hrec_crop <- crop(hunter_risk, extent(hopland))
# hrec_hunter_risk <- mask(hrec_crop, st_zm(hopland))
# 
hunter <- tm_shape(hunt_rsf) + tm_raster(style = "quantile") +
  tm_layout(aes.palette = list(seq = "YlOrRd"),
            legend.show = T, panel.show=T,
            panel.labels="Lethal Removal Risk", panel.label.size = 1,
            legend.title.color = "white", legend.outside = T,
            title.position=c("center","bottom"))
```


### Lion Risk layer

Model = Veg Type + Distance to Cropland
*(Gaynor et al. 2021)*

```{r}
#Create fake data
set.seed(42)
dummy_data2 <- data.frame(used = rbinom(30, 1, 0.4),
                         vegetation = factor(rep(c("Shrubland", "Woodland", "Grassland"), each = 10), 
                                              levels = c("Shrubland", "Woodland", "Grassland")),
                         inv_crop_dist = c(runif(30, min = 0, max = 2500))) 

#Recreate formula from Gaynor et. al 2021 and specify interactions:
mod2 <- glm(formula = used ~ vegetation + inv_crop_dist, 
                  data = dummy_data2, family = binomial(link = "logit"))

#Create a new model to update the coefficients:
mod_real2 <- mod2

#Manually create vector of REAL coefficients from Gaynor et al. 2021:
mod_real2$coefficients <-c(-5.84, -2.53, -1.02, -1.32)
```


```{r}
#Create lion risk layer from Gaynor et al. 2021 model

#create crop distance layer (as proxy for BLM distance)
crop_sc <- sdmvspecies::rescale(crop_dist)
crop_inverse = crop_sc
values(crop_inverse) <- 1 - values(crop_sc) 
plot(crop_inverse)

#stack layers
lion_stack <- brick(veg, crop_inverse) 
names(lion_stack) <- c("vegetation", "inv_crop_dist")

#convert to matrix 
lion_values <- values(lion_stack)
head(lion_values)

#create dataframe with categorical values for prediction
lion_vals <- lion_values %>% data.frame %>% 
  mutate(rid = 1:length(vegetation), nick_name = "dummy") %>% 
  merge(veg_table, all.x=TRUE) %>% #merge categorical names from before
  dplyr::select(-vegetation) %>% #drop numeric
  rename(vegetation = vegetation_class) %>% #keep factor class
  arrange(rid) #add row ID
str(lion_vals)

#predict values
lion_prediction <- predict(mod_real2, newdata = lion_vals, allow.new.levels=TRUE)

#fill raster with values
lion_risk <- lion_stack[[2]] %>% setValues(lion_prediction)

plot(lion_risk)
```

```{r}
lion <- exp(lion_risk)/max(values(exp(lion_risk)), na.rm = TRUE)

#remove dev from lion risk
lion_risk <- mask(lion, dev_mask, inverse=TRUE)

# replacing NA's by zero
lion_risk[is.na(lion_risk[])] <- 0 

plot(lion_risk)
```

```{r}
# #crop to hopland boundary to see if its accurate
# lion_crop <- crop(lion_risk, extent(hopland))
# hrec_lion_risk <- mask(lion_crop, st_zm(hopland))
# 
# lion <- tm_shape(hrec_lion_risk) + 
#           tm_raster(style = "cont", palette = "YlOrRd") + 
#           tm_shape(hopland) + tm_borders(col = "black") + 
#           tm_layout(legend.show = T, panel.show=T, panel.labels="HREC Lion Risk")
```

### HREC plots
```{r}
# tmap_arrange(hunter, lion)
```


### Models
```{r}
envt <- stack(elev, tri, veg, dev_dist, road_dist, crop_dist, hunter_risk, lion_risk)
names(envt) <- c("Elevation", "Ruggedness", "Vegetation Type", "Development Dist.",
                 "Road Dist.","Agriculture Dist", "Lethal Removal Risk", "Mountain Lion Risk")
save(envt, file = "data_objects/envt.Rdata")
```

```{r}
load(here("data_objects","envt.Rdata"))

tmlayer <- function(layer,title, color){
  tm_shape(layer) + 
    tm_raster(style = "cont", palette = "seq") +
  tm_layout(aes.palette = list(seq = color)) +
    tm_layout(legend.show = T, panel.show=T, panel.labels=title, panel.label.size = 1,
              legend.title.color = "white", legend.outside = T, title.position=c("center","bottom")) 
}

tmap_arrange(
  tmlayer(envt[[2]], "Ruggedness", "Spectral"),
  tmlayer(envt[[8]], "Mountain Lion Risk", "YlOrRd"),
  tmlayer(envt[[4]], "Development Dist.", "-YlOrRd"),
  tmlayer(envt[[5]], "Road Dist.", "-YlOrRd"), 
  tmlayer(envt[[6]], "Agriculture Dist.", "-YlGn"), 
  hunter, 
  tmlayer(envt[[3]], "Vegetation type", "Dark2"), ncol = 3) %>% 
  tmap_save('figs/Figure_S1.png')
```
