==> 3_moveHMM.Rmd <==
---
title: "moveHMM"
author: "Amy Van Scoyoc"
date: "4/5/2022"
output: github_document
---

```{r}
library(here) #reproducible working directory
library(moveHMM) #package to make HMMs
library(parallel) #package for parallel computations
```

### Import coyote dataset
```{r}
load(here("data_objects","dat_all.Rdata"))
```

### Process coyote data for HMM
```{r}
dat_processed <- prepData(dat_all, type="UTM")
plot(dat_processed, ask=F)
```

### Get zeromass parameters
```{r}
# Indices of steps of length zero
whichzero <- which(dat_processed$step == 0)
# Proportion of steps of length zero in the data set
length(whichzero)/nrow(dat_processed)
```

### Iterate to fit parameters (in parallel)
```{r}
#set seed for reproducibility
set.seed(12345)
# Create cluster of size ncores
ncores <- detectCores() - 1
#cl <- makeCluster(getOption("cl.cores", ncores)) #this wasn't working
cl <- parallel::makeCluster(2, setup_strategy = "sequential") #found this solution instead

# Export objects needed in parallelised function to cluster 
clusterExport(cl, list("dat_processed", "fitHMM"))

# Number of tries with different starting values
niter <- 25

# Create list of starting values
allPar0 <- lapply(as.list(1:niter), function(x) { 
    # Step length mean
    stepMean0 <- runif(3,
                    min = c(50, 500, 1000), 
                    max = c(100, 1000, 3000))
    # Step length standard deviation
    stepSD0 <- runif(3,
                     min = c(25, 250, 500),
                     max = c(50, 500, 1500)) 
    # Zero mass parameters 
    zeromass <- c(0.002356343,0.002356343,0.002356343)
    # Turning angle mean
    angleMean0 <- c(pi, pi/2, 0)
    # Turning angle concentration
    angleCon0 <- runif(3,
                       min = c(0.2, 0.5, 0.7),
                       max = c(0.5, 0.7, 3))
    # Return vectors of starting values
    stepPar0 <- c(stepMean0, stepSD0, zeromass)
    anglePar0 <- c(angleMean0, angleCon0)
    return(list(step = stepPar0, angle = anglePar0))
})
# Fit the niter models in parallel
allm_parallel <- parLapply(cl = cl, X = allPar0, fun = function(par0) { 
  m <- fitHMM(data = dat_processed, nbStates = 3, 
              stepPar0 = par0$step,anglePar0 = par0$angle)
  return(m) 
  })
```

### Check the best fitting model
```{r}
# Extract likelihoods of fitted models
allnllk <- unlist(lapply(allm_parallel, function(m) m$mod$minimum))
allnllk
```
 
```{r}
#get stats for likelihoods
likelihoods <- c(148653., 152963.6, 148653.3, 148653.3, 152963.6, 147916.6, 152963.6, 148653.3, 147916.6, 152963.6,152963.6, 147916.6, 148653.3, 147916.6, 152963.6, 152963.6, 147916.6, 148653.3, 152963.6, 148653.3,147916.6, 147916.6, 147916.6, 147916.6, 148653.3)

max(likelihoods)
mean(likelihoods)
sd(likelihoods)
```


```{r}
# Index of best fitting model (smallest negative log-likelihood)
whichbest <- which.min(allnllk)
# Best fitting model
mbest <- allm_parallel[[whichbest]]
mbest
plot(mbest, ask=F)
# Export as model data object
save(mbest, file = "data_objects/best_hmm.Rdata")
```

```{r}
# Assign the best model's states to the dataset 
hmm_dat <- dat_processed
hmm_dat$states <- viterbi(mbest)
# Export as csv to partition by behavioral state
write_csv(hmm_dat, "data/hmm_dat.csv")
```

```{r}
hmm_dat <- read_csv("data/hmm_dat.csv")
hmm_dat %>% group_by(states) %>% summarize(Observations = n())
```
```{r}
activity <- hmm_dat %>% 
  left_join(datekey, by = "ID") %>% #add season to check variation 
  dplyr::select(1:7, season) %>% #drop unnecessary columns
  group_by(season, ID, states) %>% 
  summarize(state_total = n()) %>% 
  ungroup() %>% 
  group_by(ID) %>% 
  mutate(total = sum(state_total), 
         percent = state_total / total * 100)

aov_rest <- aov(percent ~ as.factor(season), data = filter(activity, states == 1))
aov_forage <- aov(percent ~ as.factor(season), data = filter(activity, states == 2))
aov_travel <- aov(percent ~ as.factor(season), data = filter(activity, states == 3))

summary(aov_rest)   # significantly different
summary(aov_forage) # no difference
summary(aov_travel) # no difference

TukeyHSD(aov_rest) #3rd season % resting was significantly higher than seasons 1 and 2

ssn_act_plot <- hmm_dat %>% 
  left_join(datekey, by = "ID") %>% #add season to check variation 
  dplyr::select(1:7, season) %>% #drop unnecessary columnshmm_dat %>% 
  mutate(hour = hour(t)) %>% 
  group_by(season, ID, states, hour) %>% 
  summarize(n = n())

ggplot(ssn_act_plot, aes(x = hour, y = n, group = states, fill = factor(states))) +
    coord_polar(theta = "x", start = -.13) +
    geom_bar(stat = "identity") +
    #geom_hline(yintercept = seq(0, 500, by = 100), color = "grey80", size = 0.3) +
    scale_x_continuous(breaks = 0:24, expand = c(.002,0)) +
    scale_fill_manual(labels = c("Resting", "Foraging", "Traveling"), 
                      values = c("#E7B800", "#00afbb", "seagreen")) +
    labs(x = "Time of Day", y = "Frequency", fill = "") +
    theme_bw() +
    facet_wrap( ~ season, ncol=3)
```

```{r}
activity <- hmm_dat %>% 
  mutate(hour = hour(t)) %>% 
  group_by(ID,states, hour) %>% 
  summarize(n = n())

ggplot(activity, aes(x = hour, y = n, group = states, fill = factor(states))) +
    coord_polar(theta = "x", start = -.13) +
    geom_bar(stat = "identity") +
    #geom_hline(yintercept = seq(0, 500, by = 100), color = "grey80", size = 0.3) +
    scale_x_continuous(breaks = 0:24, expand = c(.002,0)) +
    scale_fill_manual(labels = c("Resting", "Foraging", "Traveling"), 
                      values = c("#E7B800", "#00afbb", "seagreen")) +
    labs(x = "Time of Day", y = "Frequency", fill = "") +
    theme_bw() +
    facet_wrap( ~ ID, ncol=4)
```

```{r}
hmm_dat %>% 
  group_by(ID, states) %>% 
  summarize(state_total = n()) %>% 
  ungroup() %>% 
  group_by(ID) %>% 
  mutate(total = sum(state_total), 
         percent = state_total / total * 100) %>% 
  group_by(states) %>% 
  summarize(mean_percent = mean(percent))
```
```{r}
ggplot(activity, aes(x = hour, y = n, group = states, fill = factor(states))) +
    coord_polar(theta = "x", start = -.13) +
    geom_bar(stat = "identity") +
    #geom_hline(yintercept = seq(0, 500, by = 100), color = "grey80", size = 0.3) +
    scale_x_continuous(breaks = 0:24, expand = c(.002,0)) +
    scale_fill_manual(labels = c("Resting", "Foraging", "Traveling"), 
                      values = c("#E7B800", "#00afbb", "seagreen")) +
    labs(x = "Time of Day", y = "Frequency", fill = "") +
    theme_bw() 
ggsave('figs/Figure_2.png')
```
