==> 4_full_mod.Rmd <==
---
title: "Full_mods"
author: "Amy Van Scoyoc"
date: "12/1/2022"
output: github_document
---

```{r}
library(here) #reproducible working directory
library(tidyverse) #data wrangling
library(sp) #for spatial points
library(sf) #for spatial transformations with pipes
library(lubridate) #setting timezone
library(maptools) #setting day/night times
library(adehabitatHR) #for calculating homerange KUDs
library(raster) #for importing and extracting covariates
library(lme4) #for glmer models
library(performance) #check models
library(jtools) #plot estimates
library(sjPlot)
```

### Load coyote data with HMM states, add day/night

```{r}
#import dataset
coyote <- read_csv(here("data", "hmm_dat.csv"))
#transform to geographic crs for sunriset
coordinates(coyote) <- c("x", "y")
proj4string(coyote) <- CRS("+init=epsg:26910")
coyote <- spTransform(coyote, CRS("+init=epsg:4326")) %>% as.data.frame() 
#add sunriset times
coyote <- coyote %>% 
  mutate(t = force_tz(t,tz = "America/Los_Angeles"), #declare timezone
         sunrise = sunriset(SpatialPoints(cbind(x,y),proj4string=CRS("+init=epsg:4326")), 
                                      t, direction="sunrise", POSIXct.out=TRUE)[,2],
         sunset = sunriset(SpatialPoints(cbind(x,y),proj4string=CRS("+init=epsg:4326")), 
                                      t, direction="sunset", POSIXct.out=TRUE)[,2],
         daynight = ifelse(t > sunrise & t < sunset, 1, 0), #assign tags, 1=day, 0=night 
         sex = ifelse(grepl("^F",ID),"female","male")) %>% 
  dplyr::select(ID, x, y, t, states, daynight, sex)

save(coyote, file = "data_objects/coyote.Rdata")
```

```{r}
#import covariates
load(here("data_objects","envt.Rdata"))
```

### Calculate Home Ranges (Kernel Utilization Distribution)
```{r}
#create object for calculating HRs
coyote_sp <- coyote %>% 
   dplyr::select(ID,x,y) %>% #select data 
   st_as_sf(., coords = c("x","y"), crs = "+init=epsg:4326") %>%  #make spatial object
   st_transform("+init=epsg:26910") %>% #transform back to UTMs
   as("Spatial") #transform back to sp object for KUDs

#create Kernel Utilization Distributions 
kud_all <- adehabitatHR::kernelUD(coyote_sp, h = 'href')
coyote_hr <- adehabitatHR::getverticeshr(kud_all, percent = 95)

#HR area (km2)
kernel.area(kud_all, percent = 95) %>% 
  pivot_longer(1:8,names_to = "ID", values_to = "Area (m^2)")
```

### Sample available points in Home Ranges
```{r}
coyote_full <- coyote %>% mutate(Used = 1) %>% #add column to indicate used points
   st_as_sf(., coords = c("x","y"), crs = "+init=epsg:4326") %>%  #make spatial object
   st_transform("+init=epsg:26910")  #transform back to UTMs

coyote_ids <- unique(coyote_full$ID) # Extract a list of IDs
availables <- list() #empty list to store available pts by ID

# Randomly sample available points from within the home range of each individual
for(i in 1:length(coyote_ids)){
    st_sample(st_as_sf(coyote_hr)[i,], 5*nrow(filter(coyote_full, ID == coyote_ids[i]))) %>%  #5:1 ratio
    st_sf(geometry = .) %>%
    mutate(ID = coyote_ids[i], 
           t = NA,
           states = NA,
           daynight = NA,
           sex = NA,
           Used = 0) -> availables[[i]]  
}

# Then combine individual lists into one data.frame
# Then merge with used coyote GPS points data frame
coyote_full <- availables %>% 
  do.call(rbind,.) %>% 
  rbind(coyote_full, .) %>%  #add covariates
  mutate(elev = raster::extract(envt[[1]], as(., "Spatial")) %>% scale(., center = TRUE, scale = TRUE),
         tri = raster::extract(envt[[2]], as(., "Spatial")) %>% scale(., center = TRUE, scale = TRUE),
         veg = raster::extract(envt[[3]], as(., "Spatial"), method = 'simple', factors = TRUE), 
         dev_dist = raster::extract(envt[[4]], as(., "Spatial")) %>% scale(., center = TRUE, scale = TRUE),
         road_dist = raster::extract(envt[[5]], as(., "Spatial")) %>% scale(., center = TRUE, scale = TRUE),
         crop_dist = raster::extract(envt[[6]], as(., "Spatial")) %>% scale(., center = TRUE, scale = TRUE), 
         hunter_risk = raster::extract(envt[[7]], as(., "Spatial")) %>% scale(., center = TRUE, scale = TRUE),
         lion_risk = raster::extract(envt[[8]], as(., "Spatial")) %>% scale(., center = TRUE, scale = TRUE)) %>% 
  left_join(.,envt[[3]]@data@attributes[[1]], by = c('veg' = 'ID')) %>%  #join categorical vegetation type names
  dplyr::select(-veg) %>% rename("veg" = "cover") #rename categorical veg column

# Check that we have a 5:1 ratio of used:available points
table(coyote_full$ID,coyote_full$Used)
```

```{r}
# Save data object
save(coyote_full, file = "data_objects/coyote_full.Rdata")
```

### Check covariates for colinearity
```{r}
coyote_full$ID <- as_factor(coyote_full$ID) #make sure ID is treated as a factor

coyote_cor <- coyote_full
st_geometry(coyote_cor) <- NULL

#check correlation of binary and non-binary  
full_cor <- coyote_cor %>% 
             rowid_to_column() %>% 
             mutate(veg_val = 1) %>% 
             pivot_wider(names_from = veg, values_from = veg_val, values_fill = 0) %>% 
             dplyr::select(-"NA") 
cor(full_cor[,9:17])

#remove elevation (correlated w/ crop)
#remove shrubland  (correlated w/ lion risk)

library(kableExtra)
full_cor[,c(9:17)] %>% 
  set_names("Ruggedness", "Development", "Road", 
            "Agriculture", "Lethal Removal Risk", 
            "Mountain Lion Risk", "Grassland", "Shrubland", "Woodland") %>% 
  cor() %>% 
  knitr::kable() %>% 
  kable_styling("striped", full_width = FALSE, htmltable_class = 'lightable-classic-2') %>% 
  kable_paper() %>% 
  save_kable(file = 'figs/Table_S1.png') 
```

###Full models 
```{r}
full_mod <- glmer(Used ~ tri + road_dist + dev_dist + crop_dist + hunter_risk + lion_risk + (1|ID), 
                  data = coyote_full, family=binomial(link="logit")) #drop elev (correlated w/ crop)

save(full_mod, file = "data_objects/full_mod.Rdata")
```

###Vegetation full model
```{r}
full_veg <- glmer(Used ~ veg + (1|ID), data = coyote_full, family=binomial(link="logit")) 
save(full_veg, file = "data_objects/full_veg.Rdata")
```
