==> 7_model_comparison.Rmd <==
---
title: "7_model_comparison"
author: "Amy Van Scoyoc"
date: "1/12/2023"
output: html_document
---

### Import Libraries
```{r}
library(here) #reproducibly load data 
library(kableExtra) #make basic tables
library(webshot) #make model tables
```

### Used vs. Available table
```{r}
load(here("data_objects","coyote_full.Rdata"))
load(here("data_objects","coyote_rest.Rdata"))
load(here("data_objects","coyote_forage.Rdata"))
load(here("data_objects","coyote_travel.Rdata"))
load(here("data_objects","coyote_day.Rdata"))
load(here("data_objects","coyote_night.Rdata"))

rbind(table(coyote_full$Used),
           table(coyote_rest$Used),
           table(coyote_forage$Used),
           table(coyote_travel$Used),
           table(coyote_day$Used),
           table(coyote_night$Used)) %>% 
  as.data.frame() %>% rename(Available = 1, Used = 2) %>% 
  mutate(Model = c("Full", "Rest", "Forage", "Travel", "Day", "Night")) %>% 
  dplyr::select(Model, Used, Available) %>%   
  knitr::kable() %>% 
  kable_styling("striped", full_width = FALSE, htmltable_class = 'lightable-classic-2') %>% 
  kable_paper() %>% 
  save_kable(file = 'figs/Table_S3.png') 
```

### Full model table 
```{r}
load(here("data_objects","full_mod.Rdata"))
load(here("data_objects","rest_mod.Rdata"))
load(here("data_objects","forage_mod.Rdata"))
load(here("data_objects","travel_mod.Rdata"))
load(here("data_objects","day_mod.Rdata"))
load(here("data_objects","night_mod.Rdata"))

sjPlot::tab_model(full_mod, rest_mod, forage_mod, travel_mod, day_mod, night_mod, 
          dv.labels = c("Full model","Rest","Forage","Travel", "Day", "Night"),
          pred.labels = c("Intercept", "Ruggedness", "Road distance", 
                          "Development distance", "Agriculture distance",
                          "Lethal Removal Risk", "Mountain Lion Risk"),
          show.re.var = FALSE, show.ngroups = FALSE, show.icc = FALSE,
          string.pred = "Covariates", string.ci = "CI", p.style = "stars",
          file = "figs/Table_S4.html")

# then take html file and make .png file
webshot("figs/Table_S4.html", "figs/Table_S4.png")
```

### Veg model table 
```{r}
load(here("data_objects","full_veg.Rdata"))
load(here("data_objects","rest_veg.Rdata"))
load(here("data_objects","forage_veg.Rdata"))
load(here("data_objects","travel_veg.Rdata"))
load(here("data_objects","day_veg.Rdata"))
load(here("data_objects","night_veg.Rdata"))

sjPlot::tab_model(full_veg, rest_veg, forage_veg, travel_veg, day_veg, night_veg, 
          dv.labels = c("Full model","Rest","Forage","Travel", "Day", "Night"),
          pred.labels = c("Intercept", "Shrubland", "Woodland"),
          show.re.var = FALSE, show.ngroups = FALSE, show.icc = FALSE,
          string.pred = "Covariates", string.ci = "CI", p.style = "stars", 
          file = "figs/Table_S5.html")

# then take html file and make .png file
webshot("figs/Table_S5.html", "figs/Table_S5.png")
```
